requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anthropic
anyio==3.7.1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.1.0
blinker==1.7.0
build==1.0.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chainlit
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.1.1
curl_cffi==0.6.0b9
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.8
distro==1.9.0
docstring-inheritance==2.1.2
duckduckgo_search==4.4
easydict==1.10
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi
fastapi-socketio==0.0.10
fastjsonschema==2.19.1
ffmpy==0.3.1
filelock==3.13.1
filetype==1.2.0
Flask==2.3.2
Flask-SocketIO==5.3.4
flatbuffers==23.5.26
fonttools==4.48.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.41
google-auth==2.27.0
googleapis-common-protos==1.62.0
gradio
gradio_client
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.29.1
ipython==8.13.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jq==1.6.0
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
kubernetes==29.0.0
langchain
langchain-community
langchain-core
langchain-experimental
langchain-openai
langsmith
Lazify==0.4.0
literalai==0.0.103
loguru==0.7.2
lxml==5.1.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.15.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
notebook==7.0.7
notebook_shim==0.2.3
notion-client==2.2.1
numexpr==2.8.6
numpy==1.24.3
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.11.1
opencv-python==4.9.0.80
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pinecone-client==3.0.3
platformdirs==3.11.0
plotly==5.18.0
posthog==3.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic
pydantic_core==2.16.2
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==4.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-json-logger==2.0.7
python-multipart==0.0.6
python-socketio==5.11.1
pytz==2024.1
PyYAML==6.0
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruff==0.2.1
safetensors==0.4.2
semantic-version==2.10.0
Send2Trash==1.8.2
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette
streamlit==1.28.1
streamlit-feedback==0.1.3
sympy==1.12
syncer==2.0.3
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.0
text-generation==0.6.1
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.15.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.37.2
trubrics==1.6.2
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uptrace==1.22.0
uri-template==1.3.0
urllib3==2.2.0
uvicorn==0.23.2
uvloop==0.19.0
validators==0.22.0
virtualenv==20.23.0
watchdog==4.0.0
watchfiles==0.20.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==11.0.3
Werkzeug==2.3.6
widgetsnbextension==4.0.9
Wikipedia-API==0.6.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
